requirements.txt
asgiref==3.4.1
Django==3.2.12
django-cors-headers==3.10.1
django-ninja==0.17.0
mysqlclient==2.1.0
orjson==3.6.1
pydantic==1.9.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
